third_party/WebKit/LayoutTests/http/tests/websocket/long-invalid-header-expected.txt

CONSOLE ERROR: line 19: WebSocket connection to 'ws://127.0.0.1:8880/long-invalid-header' failed: Error during WebSocket handshake: 'Upgrade' header is missing
Make sure WebSocket gives errors on long invalid upgrade header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
WebSocket is closed
PASS successfullyParsed is true

TEST COMPLETE
